+++ mitra_api/src/mastodon_api/favourites/views.rs
@@ -15,7 +15,7 @@ use mitra_models::{
         get_database_client,
         DatabaseConnectionPool,
     },
-    reactions::queries::get_liked_posts,
+    reactions::queries::get_reactions,
 };
 
 use crate::{
@@ -43,19 +43,20 @@ async fn get_favourites_view(
 ) -> Result<HttpResponse, MastodonError> {
     let db_client = &**get_database_client(&db_pool).await?;
     let current_user = get_current_user(db_client, auth.token()).await?;
-    let liked_posts = get_liked_posts(
+    let reactions = get_reactions(
         db_client,
         current_user.id,
+        !query_params.all_reactions,
         query_params.max_id,
         query_params.limit.inner(),
     ).await?;
     let base_url = get_request_base_url(connection_info);
     let authority = Authority::from(&config.instance());
     let media_server = ClientMediaServer::new(&config, &base_url);
-    let maybe_last_id = get_last_item(&liked_posts, &query_params.limit)
-        .map(|liked_post| liked_post.reaction_id);
-    let posts = liked_posts.into_iter()
-        .map(|liked_post| liked_post.post)
+    let maybe_last_id = get_last_item(&reactions, &query_params.limit)
+        .map(|reaction| reaction.reaction_id);
+    let posts = reactions.into_iter()
+        .map(|reaction| reaction.post)
         .collect();
     let statuses = build_status_list(
         db_client,